Ya funciona el filtro
