+++ src/view.js
@@ -1,13 +1,12 @@
 // estas funciones son de ejemplo
-import data2 from "./data/ghibli/ghibli.js";
-//console.log(data.films);
+
 export const renderItems = (data) => {
-  const films = data.films;
+  const films = data;
   let titlesHTML = "";
 
   const ul = document.createElement("ul");
   ul.id = "container";
-
+console.log(data);
   //identificacmos los datos de las peliculas
   films.forEach((movie) => {
     titlesHTML += `
@@ -32,26 +31,7 @@ export const renderItems = (data) => {
   return ul;
 };
 
-window.addEventListener("DOMContentLoaded", () => {
-  //console.log(data2.films);
-  const masInfo = document.querySelectorAll(".info");
-  //console.log(buttons);
-  let arrInfo = [];
-  masInfo.forEach((eventClick) => {
-    eventClick.addEventListener("click", (e) => {
-      // console.log("desde aqui", e.target.id);
-      data2.films.filter((movie) => {
-        if (movie.id === e.target.id) {
-          // console.log(movie.people);
-          arrInfo = movie;
-        }
-      });
-      //console.log(arrInfo);
-      localStorage.setItem("movieInfoLocalStorage", JSON.stringify(arrInfo));
-      window.location.href = "info.html";
-    });
-  });
-});
+
 
 export const renderItems2 = () => {
   const info = JSON.parse(localStorage.getItem("movieInfoLocalStorage"));
